Add test to detect excessive memory growth
This test is specifically focused on detecting excessive memory
growth while iterating over a simple set (an obvious failure).

Admittedly, this test hasn't been verified on multiple
environments.  In the future it would be good to set up some
automated testing to run this on a more diverse set of platforms.

This test should be removed if found unreliable.