Rename 10_rt_58389.t
The 10_rt_58389.t test name doesn't contain context about what
it's testing.  Even though rt.cpan.org most likely won't go away
in the short term, we (as a community) should consider not
referencing parts within the Perl ecosystem that may not be around
in the next 10, 5, or even 1 year.

Developers new to the Perl ecosystem might not have any context to
what RT is in relation to Perl's history.  Even using Github issue
numbers in commit messages is probably a bad idea for the same
reason.